# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Events API - Luma-Inspired Event Management",
        "description": "Open-source headless Laravel event management API from Maak. Features organizations, events, ticket types, registrations, QR check-ins, waitlist management, and team member invitations.",
        "contact": {
            "name": "Maak",
            "url": "https://github.com/maakwizardry/events"
        },
        "version": "1.1.0"
    },
    "servers": [
        {
            "url": "http://events.test/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "PHP Built-in Server"
        }
    ],
    "paths": {
...
        "/organizations": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "List user organizations",
                "description": "Get all organizations where the authenticated user is a member",
                "operationId": "676bef8dfde354591a3d9b90ecc9e0ff",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of organizations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "uuid": {
                                                        "type": "string",
                                                        "example": "550e8400-e29b-41d4-a716-446655440000"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Tech Community"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "tech-community"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "logo_url": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "website_url": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "members_count": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "events_count": {
                                                        "type": "integer",
                                                        "example": 12
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Create a new organization",
                "description": "Create a new organization and become its owner",
                "operationId": "6951556bcf75a635eee4f116b012baf2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Tech Community"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "tech-community"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A community for tech enthusiasts"
                                    },
                                    "logo_url": {
                                        "type": "string",
                                        "format": "url",
                                        "nullable": true
                                    },
                                    "website_url": {
                                        "type": "string",
                                        "format": "url",
                                        "nullable": true
                                    },
                                    "primary_color": {
                                        "type": "string",
                                        "example": "#3B82F6"
                                    },
                                    "secondary_color": {
                                        "type": "string",
                                        "example": "#1E40AF"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Organization created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/organizations/{organization}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Get organization details",
                "description": "Get details of a specific organization",
                "operationId": "6cf169df05a11d2bc77f98141d608788",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "description": "Organization UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organization details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "404": {
                        "description": "Organization not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Update organization",
                "description": "Update organization details (owner/admin only)",
                "operationId": "0b609056c068cc14922a8074b916f1db",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "description": "Organization UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "logo_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "website_url": {
                                        "type": "string",
                                        "format": "url"
                                    },
                                    "primary_color": {
                                        "type": "string"
                                    },
                                    "secondary_color": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Organization updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not authorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Delete organization",
                "description": "Delete an organization (owner only)",
                "operationId": "3399797ce39a0c2785d248fd1978e167",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "description": "Organization UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organization deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Organization deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not authorized - owner only"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use the token from POST /auth/login or POST /auth/register",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}